aggregator/requirements.txt
# Kafka consumer
aiokafka==0.8.1

# ClickHouse database
clickhouse-driver==0.2.5

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3
tenacity==8.2.2
json-logging==1.3.0
pydantic==1.10.7

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0